Updated to new UI in internal pages
