+++ client/flutter/lib/listViewPage.dart
@@ -1,0 +1,63 @@
+import 'package:flutter/material.dart';
+
+class ListOfItemsPage extends StatelessWidget {
+  final List<ListItem> listOfItems;
+  ListOfItemsPage(this.listOfItems);
+
+  @override
+  Widget build(BuildContext context) {
+    return Material(
+      child: CustomScrollView(
+        slivers: [
+          SliverAppBar(
+            flexibleSpace: Container(),
+            expandedHeight: 200,
+          ),
+          SliverList(
+            delegate: SliverChildListDelegate(this.listOfItems),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class ListItem extends StatelessWidget {
+  final Widget titleWidget;
+  final String message;
+
+  ListItem({this.titleWidget, this.message = ""});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Column(
+        children: <Widget>[
+          this.titleWidget ?? Divider(),
+          Text(
+            this.message,
+            textScaleFactor: 1.5,
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class EmojiHeader extends StatelessWidget {
+  EmojiHeader(this.emoji);
+
+  final String emoji;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text(
+        this.emoji,
+        textScaleFactor: 6,
+      ),
+    );
+  }
+}
