+++ client/flutter/pubspec.yaml
@@ -13,6 +13,8 @@ dependencies:
   share: ^0.6.3+6
   url_launcher: ^5.4.2
   page_view_indicator: ^2.0.4
+  flutter_staggered_grid_view: ^0.3.0
+
   intl: ^0.16.0
   flutter_localizations:
     sdk: flutter
